requirements.txt
PyYAML>=3.11
jira>=1.0.7
